requirements.txt
flask
flask-smorest 
python-dotenv
sqlalchemy
flask-sqlalchemy
marshmallow
requests
rq
redis
